Split drafts to their dedicated folders
